# docs/swagger.yaml
definitions:
...
  models.DeleteResponse:
    properties:
      message:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.PhotoCreateInput:
    properties:
      caption:
        type: string
      photoURL:
        type: string
      title:
        type: string
      userID:
        type: integer
    type: object
  models.PhotoCreateOutput:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.PhotoGetOutput:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.UserRegisterOutput'
    type: object
  models.PhotoUpdateInput:
    properties:
      caption:
        type: string
      id:
        type: integer
      photoURL:
        type: string
      title:
        type: string
      userID:
        type: integer
    type: object
  models.PhotoUpdateOutput:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  models.UserRegisterOutput:
    properties:
      age:
        type: integer
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/photos:
    get:
      description: Get all photos
      parameters:
      - description: 'format: Bearer token-here'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PhotoGetOutput'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all photos
      tags:
      - photos
    post:
      consumes:
      - multipart/form-data
      description: Create photos
      parameters:
      - description: create photo
        in: body
        name: models.PhotoCreateInput
        required: true
        schema:
          $ref: '#/definitions/models.PhotoCreateInput'
      - description: 'format: Bearer token-here'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.PhotoCreateOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create photos
      tags:
      - photos
  /api/v1/photos/:photoId:
    delete:
      description: Delete photo
      parameters:
      - description: delete photo by id
        in: path
        name: photoId
        required: true
        type: string
      - description: 'format: Bearer token-here'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.DeleteResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete photo
      tags:
      - photos
    get:
      description: Get one photo by id
      parameters:
      - description: get photo by id
        in: path
        name: photoId
        required: true
        type: string
      - description: 'format: Bearer token-here'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PhotoGetOutput'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get one photo by id
      tags:
      - photos
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update photo
      parameters:
      - description: update photo by id
        in: path
        name: photoId
        required: true
        type: string
      - description: update photo
        in: body
        name: models.PhotoUpdateInput
        required: true
        schema:
          $ref: '#/definitions/models.PhotoUpdateInput'
      - description: 'format: Bearer token-here'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PhotoUpdateOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update photo
      tags:
      - photos
...
swagger: "2.0"